+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Fixed crashes when checking some reactive expressions in markup
 - Fixed error locations on variable declarations
 - Fixed misleading error on compile
+- Fixed function type display
 
 ## [0.1.0] - 2026-08-29